Server_code/app/src/requirements.txt
anyio==3.7.0
async-generator==1.10
attrs==23.1.0
bcrypt==4.0.1
click==8.1.3
dnspython==2.3.0
email-validator==2.0.0.post2
exceptiongroup==1.1.1
fastapi==0.95.2
fastapi-login==1.9.0
h11==0.14.0
idna==3.4
mysql-connector-python==8.0.33
outcome==1.2.0
passlib==1.7.4
protobuf==3.20.3
pydantic==1.10.8
PyJWT==2.7.0
python-dotenv==1.0.0
python-multipart==0.0.6
sniffio==1.3.0
sortedcontainers==2.4.0
starlette==0.27.0
trio==0.21.0
typing_extensions==4.6.2
uvicorn==0.22.0
